rm token helper
