+++ deployment/virtual-waiting-room-on-aws.json
@@ -3471,9 +3471,6 @@
                     },
                     {
                         "Ref": "AssignQueueRole"
-                    },
-                    {
-                        "Ref": "GenerateTokenRole"
                     }
                 ]
             }
